refactor: 구매 페이지 폴더 구조 변경
