+++ src/app/purchase/_component/AuctionBanner.tsx
