+++ src/app/purchase/_component/AuctionRanking.tsx
